# [SDPA] [MPS] Fixes regression in 2.8.0 for scaled_dot_product_attention using mps