--- a/aten/src/ATen/native/mps/kernels/Attention.metal
+++ b/aten/src/ATen/native/mps/kernels/Attention.metal
@@ -14,12 +14,13 @@ template <typename T, int D, int V = D>
     device T* out [[buffer(3)]],
     const constant uint& gqa_factor [[buffer(4)]],
     const constant uint& N [[buffer(5)]],
-    const constant uint2& k_head_seq_stride [[buffer(6)]],
-    const constant uint2& v_head_seq_stride [[buffer(7)]],
-    const constant float& scale [[buffer(8)]],
-    const device bool* mask [[buffer(9)]],
-    const constant uint3& mask_strides [[buffer(10)]],
-    const constant bool& has_mask [[buffer(11)]],
+    const constant uint2& q_head_seq_stride [[buffer(6)]],
+    const constant uint2& k_head_seq_stride [[buffer(7)]],
+    const constant uint2& v_head_seq_stride [[buffer(8)]],
+    const constant float& scale [[buffer(9)]],
+    const device bool* mask [[buffer(10)]],
+    const constant uint3& mask_strides [[buffer(11)]],
+    const constant bool& has_mask [[buffer(12)]],
     uint3 tid [[threadgroup_position_in_grid]],
     uint3 tpg [[threadgroups_per_grid]],
     uint simd_gid [[simdgroup_index_in_threadgroup]],
@@ -28,6 +29,8 @@ template <typename T, int D, int V = D>
   constexpr uint BD = 32;
   constexpr uint qk_per_thread = D / BD;
   constexpr uint v_per_thread = V / BD;
+  const uint q_head_stride = q_head_seq_stride.x;
+  const uint q_seq_stride = q_head_seq_stride.y;
   const uint k_head_stride = k_head_seq_stride.x;
   const uint k_seq_stride = k_head_seq_stride.y;
   const uint v_head_stride = v_head_seq_stride.x;
@@ -54,9 +57,8 @@ template <typename T, int D, int V = D>
   const int kv_head_idx = head_idx / gqa_factor;
   const int Q = tpg.y;
   const int group_offset = head_idx * Q + q_seq_idx;
-  const int q_offset = group_offset;
   const int o_offset = group_offset;
-  queries += q_offset * D + simd_lid * qk_per_thread;
+  queries += head_idx * q_head_stride + q_seq_idx * q_seq_stride + simd_lid * qk_per_thread;
   keys += kv_head_idx * k_head_stride + simd_gid * k_seq_stride +
       simd_lid * qk_per_thread;
   values += kv_head_idx * v_head_stride + simd_gid * v_seq_stride +
@@ -156,12 +158,13 @@ template <typename T, int D, int V = D>
     device float* maxs [[buffer(5)]],
     const constant uint& gqa_factor [[buffer(6)]],
     const constant uint& N [[buffer(7)]],
-    const constant uint2& k_head_seq_stride [[buffer(8)]],
-    const constant uint2& v_head_seq_stride [[buffer(9)]],
-    const constant float& scale [[buffer(10)]],
-    const device bool* mask [[buffer(11)]],
-    const constant uint3& mask_strides [[buffer(12)]],
-    const constant bool& has_mask [[buffer(13)]],
+    const constant uint2& q_head_seq_stride [[buffer(8)]],
+    const constant uint2& k_head_seq_stride [[buffer(9)]],
+    const constant uint2& v_head_seq_stride [[buffer(10)]],
+    const constant float& scale [[buffer(11)]],
+    const device bool* mask [[buffer(12)]],
+    const constant uint3& mask_strides [[buffer(13)]],
+    const constant bool& has_mask [[buffer(14)]],
     uint3 tid [[threadgroup_position_in_grid]],
     uint3 tpg [[threadgroups_per_grid]],
     uint simd_gid [[simdgroup_index_in_threadgroup]],
@@ -170,6 +173,8 @@ template <typename T, int D, int V = D>
   constexpr int BD = 32;
   constexpr int qk_per_thread = D / BD;
   constexpr int v_per_thread = V / BD;
+  const int q_head_stride = q_head_seq_stride.x;
+  const int q_seq_stride = q_head_seq_stride.y;
   const int k_head_stride = k_head_seq_stride.x;
   const int k_seq_stride = k_head_seq_stride.y;
   const int v_head_stride = v_head_seq_stride.x;
@@ -196,10 +201,9 @@ template <typename T, int D, int V = D>
   const int head_idx = tid.x;
   const int q_seq_idx = tid.y;
   const int o_offset = head_idx * tpg.y + q_seq_idx;
-  const int q_offset = o_offset;
   const int kv_head_idx = head_idx / gqa_factor;
 
-  queries += q_offset * D + simd_lid * qk_per_thread;
+  queries += head_idx * q_head_stride + q_seq_idx * q_seq_stride + simd_lid * qk_per_thread;
   keys += kv_head_idx * k_head_stride +
       (block_idx * BN + simd_gid) * k_seq_stride + simd_lid * qk_per_thread;
   values += kv_head_idx * v_head_stride +
@@ -530,12 +534,13 @@ kernel void attention(
       device DTYPE* out [[buffer(3)]],                       \
       const constant uint& gqa_factor [[buffer(4)]],         \
       const constant uint& N [[buffer(5)]],                  \
-      const constant uint2& k_head_seq_stride [[buffer(6)]], \
-      const constant uint2& v_head_seq_stride [[buffer(7)]], \
-      const constant float& scale [[buffer(8)]],             \
-      const device bool* mask [[buffer(9)]],                 \
-      const constant uint3& mask_strides [[buffer(10)]],     \
-      const constant bool& has_mask [[buffer(11)]],          \
+      const constant uint2& q_head_seq_stride [[buffer(6)]], \
+      const constant uint2& k_head_seq_stride [[buffer(7)]], \
+      const constant uint2& v_head_seq_stride [[buffer(8)]], \
+      const constant float& scale [[buffer(9)]],             \
+      const device bool* mask [[buffer(10)]],                \
+      const constant uint3& mask_strides [[buffer(11)]],     \
+      const constant bool& has_mask [[buffer(12)]],          \
       uint3 tid [[threadgroup_position_in_grid]],            \
       uint3 tpg [[threadgroups_per_grid]],                   \
       uint simd_gid [[simdgroup_index_in_threadgroup]],      \
@@ -553,12 +558,13 @@ kernel void attention(
       device float* maxs [[buffer(5)]],                           \
       const constant uint& gqa_factor [[buffer(6)]],              \
       const constant uint& N [[buffer(7)]],                       \
-      const constant uint2& k_head_seq_stride [[buffer(8)]],      \
-      const constant uint2& v_head_seq_stride [[buffer(9)]],      \
-      const constant float& scale [[buffer(10)]],                 \
-      const device bool* mask [[buffer(11)]],                     \
-      const constant uint3& mask_strides [[buffer(12)]],          \
-      const constant bool& has_mask [[buffer(13)]],               \
+      const constant uint2& q_head_seq_stride [[buffer(8)]],      \
+      const constant uint2& k_head_seq_stride [[buffer(9)]],      \
+      const constant uint2& v_head_seq_stride [[buffer(10)]],     \
+      const constant float& scale [[buffer(11)]],                 \
+      const device bool* mask [[buffer(12)]],                     \
+      const constant uint3& mask_strides [[buffer(13)]],          \
+      const constant bool& has_mask [[buffer(14)]],               \
       uint3 tid [[threadgroup_position_in_grid]],                 \
       uint3 tpg [[threadgroups_per_grid]],                        \
       uint simd_gid [[simdgroup_index_in_threadgroup]],           \