@Vismai-Khanderao 2025-09-29
Thanks for the review @malfet 

> LGTM, though I wonder what's the perf implications of this change

Results of running the benchmarking script from the original PR is included in the description. The times seem within margin of error, even after addressing @Isalia20's suggestions. But a dedicated benchmark might be useful.